ENH: air density correction factor to V_D (#142)
This PR adds an automated air density correction factor following
Heymsfield et al. (JAS, 2007), which is implemented in various microphysics
schemes such as MG, MG2, and P3. The correction is applied based on the
reference air density, which is hydrometeor class-dependent, and was set
using the information from the relevant literature.
Note that this PR also adds `Model` class method to calculate air density during init
to enable smooth processing of the microphysics approach.